:recycle: 검색 결과에 사용되는 server action 함수들을 product action으로 옮겼어요
